+++ Changelog.md
@@ -5,6 +5,7 @@
   ungradeable (#4331)
 - Fixed bug where criteria were not expanded for grading (to both Admins and TAs) (#4380)
 - Fixed bug where annotations were not removed when switching between PDF submission files (#4387)
+- Fixed bug where annotations disappeared on window resize (#4387)
 
 ## [v1.8.4]
 - Fixed bug where test output was not being properly hidden from students (#4379)